refactor: replaced main.tsx to index.tsx
