+++ src/App.tsx
@@ -1,4 +1,4 @@
-import { NavigationBar } from './components/NavigationBar/main';
+import { NavigationBar } from './components/NavigationBar';
 
 function App() {
   return (
